Changed namespaces, project names, and target framework versions
