+++ src/Warframe.World/Models/ConclaveChallenge.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace WorldState.Data.Models
+namespace Warframe.World.Models
 {
     public class ConclaveChallenge
     {
