+++ src/Warframe.World/Models/SyndicateMission.cs
@@ -1,9 +1,9 @@
-﻿using Newtonsoft.Json;
-using System;
+﻿using System;
+using Newtonsoft.Json;
 
-namespace WorldState.Data.Models
+namespace Warframe.World.Models
 {
-    public class Alert
+    public class SyndicateMission
     {
         [JsonProperty]
         public string Id { get; private set; }
@@ -14,17 +14,17 @@ namespace WorldState.Data.Models
         [JsonProperty("expiry")]
         public DateTimeOffset ExpiresAt { get; private set; }
 
-        [JsonProperty]
-        public AlertMission Mission { get; private set; }
-
         [JsonProperty("active")]
         public bool IsActive { get; private set; }
 
-        [JsonProperty("expired")]
-        public bool HasExpired { get; private set; }
+        [JsonProperty]
+        public string Syndicate { get; private set; }
 
         [JsonProperty]
-        public string[] RewardTypes { get; private set; }
+        public string[] Nodes { get; private set; }
+
+        [JsonProperty("jobs")]
+        public Bounty[] Bounties { get; private set; }
 
         [JsonProperty("eta")]
         public string TimeRemaining { get; private set; }
