+++ src/Warframe/Warframe.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.6</TargetFramework>
+    <RootNamespace>Warframe</RootNamespace>
+    <AssemblyName>Warframe</AssemblyName>
+  </PropertyGroup>
+
+</Project>
