--- src/WorldState/WorldState.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.6</TargetFramework>
-    <RootNamespace>WorldState</RootNamespace>
-    <AssemblyName>WorldState</AssemblyName>
-  </PropertyGroup>
-
-</Project>
